refactor: 코드 정리 및 store subscribe handler 설정
