+++ src/controller/controlIndex.js
@@ -1,4 +1,3 @@
-import { fieldClick, fieldXScroll } from "./Components/field.js";
 import { pageButton } from "./Components/pageButton.js";
 import { pressFilterTabs } from "./Components/pressFilterTabs.js";
 import { viewButton } from "./Components/viewButton.js";
@@ -10,7 +9,4 @@ export function controller() {
   viewButton();
   //탭버튼
   pressFilterTabs();
-  //
-  fieldClick();
-  // fieldXScroll();
 }
