+++ src/view/field.js
@@ -3,6 +3,7 @@ import { VIEW, LIST_PAGE } from "../model/global.js";
 import { store } from "../model/store.js";
 import { news_data } from "./grid.js";
 import { category, eachCategoryLength, setNews } from "./list.js";
+import { fieldClick, fieldXScroll } from "../controller/Components/field.js";
 
 export function renderTabs(news) {
   let CATEGORY = category;
@@ -40,6 +41,10 @@ export function renderTabs(news) {
     const news_list_wrap = document.querySelector("main .news-list-wrap");
     news_list_wrap.appendChild(category_section_tab);
   }
+
+  //탭 이벤트 등록
+  fieldClick();
+  fieldXScroll();
 }
 export function updateField() {
   if (VIEW.tab === "entire") {
@@ -48,6 +53,7 @@ export function updateField() {
     progressTabNumber.querySelector(".present").innerHTML = `${LIST_PAGE.page + 1} / `;
   }
 }
+// store.subscribe(updateCategory);
 export function updateCategory() {
   const categoryNews = news_data.filter((press) => press.category === category[LIST_PAGE.category]);
   setNews(VIEW.tab === "entire" ? shuffle_press(categoryNews) : store.getSubscribe());
